+++ arduino/builder/sketch.go
@@ -18,12 +18,19 @@ package builder
 import (
 	"bytes"
 	"fmt"
+	"io"
 	"regexp"
 
+	"github.com/arduino/arduino-cli/arduino/builder/compilation"
 	"github.com/arduino/arduino-cli/arduino/builder/cpp"
+	"github.com/arduino/arduino-cli/arduino/builder/progress"
+	"github.com/arduino/arduino-cli/arduino/builder/utils"
 	"github.com/arduino/arduino-cli/arduino/sketch"
 	"github.com/arduino/arduino-cli/i18n"
+	f "github.com/arduino/arduino-cli/internal/algorithms"
+	rpc "github.com/arduino/arduino-cli/rpc/cc/arduino/cli/commands/v1"
 	"github.com/arduino/go-paths-helper"
+	"github.com/arduino/go-properties-orderedmap"
 
 	"github.com/pkg/errors"
 )
@@ -170,4 +177,65 @@ func writeIfDifferent(source []byte, destPath *paths.Path) error {
 
 	// Source and destination are the same, don't write anything
 	return nil
+}
+
+func SketchBuilder(
+	sketchBuildPath *paths.Path,
+	buildProperties *properties.Map,
+	includesFolders paths.PathList,
+	onlyUpdateCompilationDatabase, verbose bool,
+	compilationDatabase *compilation.Database,
+	jobs int,
+	warningsLevel string,
+	stdoutWriter, stderrWriter io.Writer,
+	verboseInfoFn func(msg string),
+	verboseStdoutFn, verboseStderrFn func(data []byte),
+	progress *progress.Struct, progressCB rpc.TaskProgressCB,
+) (paths.PathList, error) {
+	includes := f.Map(includesFolders.AsStrings(), cpp.WrapWithHyphenI)
+
+	if err := sketchBuildPath.MkdirAll(); err != nil {
+		return nil, errors.WithStack(err)
+	}
+
+	sketchObjectFiles, err := utils.CompileFiles(
+		sketchBuildPath, sketchBuildPath, buildProperties, includes,
+		onlyUpdateCompilationDatabase,
+		compilationDatabase,
+		jobs,
+		verbose,
+		warningsLevel,
+		stdoutWriter, stderrWriter,
+		verboseInfoFn,
+		verboseStdoutFn,
+		verboseStderrFn,
+		progress, progressCB,
+	)
+	if err != nil {
+		return nil, errors.WithStack(err)
+	}
+
+	// The "src/" subdirectory of a sketch is compiled recursively
+	sketchSrcPath := sketchBuildPath.Join("src")
+	if sketchSrcPath.IsDir() {
+		srcObjectFiles, err := utils.CompileFilesRecursive(
+			sketchSrcPath, sketchSrcPath, buildProperties, includes,
+			onlyUpdateCompilationDatabase,
+			compilationDatabase,
+			jobs,
+			verbose,
+			warningsLevel,
+			stdoutWriter, stderrWriter,
+			verboseInfoFn,
+			verboseStdoutFn,
+			verboseStderrFn,
+			progress, progressCB,
+		)
+		if err != nil {
+			return nil, errors.WithStack(err)
+		}
+		sketchObjectFiles.AddAll(srcObjectFiles)
+	}
+
+	return sketchObjectFiles, nil
 }
